Finish implementing model and trainer, tested

+++ trainer/dataset.py
@@ -11,9 +11,9 @@ class RoadSceneDataset(torch.utils.data.Dataset):
             transform=torchvision.transforms.ToTensor(),
     ):
         # each image is stored as np array (32,32) [0,255], the dataset is stored as numpy array (N, 32, 32), (N), (N)
-        self.images = images
-        self.offsets = offsets
-        self.angles = angles
+        self.images = images.astype(np.uint8)
+        self.offsets = offsets.astype(np.float32)
+        self.angles = angles.astype(np.float32)
         self.transform = transform # Totensor will transform (H,W,C) np array [0,255] into (C, H, W) tensor [0,1.0]
         assert self.images.shape[0] == self.offsets.shape[0], "Row mismatch"
         assert self.offsets.shape[0] == self.angles.shape[0], "Row mismatch"
